Fixing

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	long long int num = 612852475143;
+	long int num = 612852475143;
 	int max;
 	for (max = 1; max < num; max++)
 	{
